Rename `rustc_query_impl/error.rs` to `diagnostics.rs`
