+++ compiler/rustc_query_impl/src/diagnostics.rs
