+++ compiler/rustc_query_impl/src/job.rs
@@ -421,7 +421,7 @@ pub(crate) fn create_cycle_error<'tcx>(
 
     let mut cycle_stack = Vec::new();
 
-    use crate::error::StackCount;
+    use crate::diagnostics::StackCount;
     let stack_bottom = frames[0].tagged_key.catch_description(tcx);
     let stack_count = if frames.len() == 1 {
         StackCount::Single { stack_bottom: stack_bottom.clone() }
@@ -433,7 +433,7 @@ pub(crate) fn create_cycle_error<'tcx>(
     for i in 1..frames.len() {
         let frame = &frames[i];
         let span = frame.tagged_key.catch_default_span(tcx, frames[(i + 1) % frames.len()].span);
-        cycle_stack.push(crate::error::CycleStack {
+        cycle_stack.push(crate::diagnostics::CycleStack {
             span: if span == prev { DUMMY_SP } else { span },
             desc: frame.tagged_key.catch_description(tcx),
         });
@@ -442,7 +442,7 @@ pub(crate) fn create_cycle_error<'tcx>(
 
     let cycle_usage = usage.as_ref().map(|usage| {
         let cycle_span = usage.tagged_key.catch_default_span(tcx, usage.span);
-        crate::error::CycleUsage {
+        crate::diagnostics::CycleUsage {
             span: if cycle_span != span { cycle_span } else { DUMMY_SP },
             usage: usage.tagged_key.catch_description(tcx),
         }
@@ -464,9 +464,9 @@ pub(crate) fn create_cycle_error<'tcx>(
 
     let alias = if !nested {
         if is_all_def_kind(DefKind::TyAlias) {
-            Some(crate::error::Alias::Ty)
+            Some(crate::diagnostics::Alias::Ty)
         } else if is_all_def_kind(DefKind::TraitAlias) {
-            Some(crate::error::Alias::Trait)
+            Some(crate::diagnostics::Alias::Trait)
         } else {
             None
         }
@@ -475,16 +475,16 @@ pub(crate) fn create_cycle_error<'tcx>(
     };
 
     if nested {
-        tcx.sess.dcx().create_err(crate::error::NestedCycle {
+        tcx.sess.dcx().create_err(crate::diagnostics::NestedCycle {
             span,
             cycle_stack,
-            stack_bottom: crate::error::NestedCycleBottom { stack_bottom },
+            stack_bottom: crate::diagnostics::NestedCycleBottom { stack_bottom },
             cycle_usage,
             stack_count,
             note_span: (),
         })
     } else {
-        tcx.sess.dcx().create_err(crate::error::Cycle {
+        tcx.sess.dcx().create_err(crate::diagnostics::Cycle {
             span,
             cycle_stack,
             stack_bottom,
